delete Admin Service
