--- src/main/java/de/willbeedone/backend/service/interfaces/AdminService.java
@@ -1,10 +1,0 @@
-package de.willbeedone.backend.service.interfaces;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface AdminService {
-
-    boolean blockUserByEmail(String email);
-
-}
